+++ test/Helper/Navigation/AbstractTest.php
@@ -86,14 +86,6 @@ abstract class AbstractTest extends \PHPUnit_Framework_TestCase
      */
     protected function setUp()
     {
-        if (! class_exists(ServiceManagerConfig::class)) {
-            $this->markTestSkipped(
-                'Skipping zend-mvc-related tests until that component is updated '
-                . 'to be forwards-compatible with zend-eventmanager, zend-stdlib, '
-                . 'and zend-servicemanager v3.'
-            );
-        }
-
         $cwd = __DIR__;
 
         // read navigation config